# openapi.yaml
openapi: 3.0.3
info:
  title: Gamified Evidence-Based Nursing API
  version: 1.0.0
  description: OpenAPI spec generated from Spring controller mappings in this repository.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/rooms/{roomId}/status:
    get:
      tags: [Room]
      operationId: getRoomStatus
      parameters:
        - $ref: '#/components/parameters/roomIdPath'
        - $ref: '#/components/parameters/teamIdQuery'
      responses:
        '200':
          description: Room status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/rooms/{roomId}/questions:
    get:
      tags: [Room]
      operationId: getRoomQuestions
      parameters:
        - $ref: '#/components/parameters/roomIdPath'
        - $ref: '#/components/parameters/teamIdQuery'
      responses:
        '200':
          description: Questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuestionDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/rooms/{roomId}/submit-answer:
    post:
      tags: [Room]
      operationId: submitAnswer
      parameters:
        - $ref: '#/components/parameters/roomIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerRequest'
      responses:
        '200':
          description: Submission result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitAnswerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/rooms/{roomId}/questions/{questionId}/hint:
    post:
      tags: [Room]
      operationId: getHint
      parameters:
        - $ref: '#/components/parameters/roomIdPath'
        - $ref: '#/components/parameters/questionIdPath'
        - $ref: '#/components/parameters/teamIdQuery'
      responses:
        '200':
          description: Hint text
          content:
            text/plain:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/rooms/{roomId}/questions/{questionId}/skip:
    post:
      tags: [Room]
      operationId: skipQuestion
      parameters:
        - $ref: '#/components/parameters/roomIdPath'
        - $ref: '#/components/parameters/questionIdPath'
        - $ref: '#/components/parameters/teamIdQuery'
      responses:
        '200':
          description: Updated room status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/rooms/room-of-knowledge:
    get:
      tags: [Room]
      operationId: getRoomOfKnowledgeQuestionList
      parameters:
        - name: gameId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: mission
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Mission'
        - name: password
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of Room of Knowledge questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomOfKnowledgeQuestionDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/completion/teams/{teamId}/status:
    get:
      tags: [Completion]
      operationId: checkTeamCompletion
      parameters:
        - $ref: '#/components/parameters/teamIdPath'
      responses:
        '200':
          description: Completion status
          content:
            application/json:
              schema:
                type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/completion/games/{gameId}/check:
    get:
      tags: [Completion]
      operationId: checkGameCompletion
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Game result if complete, otherwise null
          content:
            application/json:
              schema:
                nullable: true
                allOf:
                  - $ref: '#/components/schemas/GameResultDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/completion/games/{gameId}/results:
    get:
      tags: [Completion]
      operationId: getGameResults
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Final game results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResultDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/completion/games/{gameId}/force-end:
    post:
      tags: [Completion]
      operationId: forceEndGame
      description: Requires ADMIN role.
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Forced final result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResultDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/completion/games/{gameId}/winner:
    get:
      tags: [Completion]
      operationId: getWinner
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Winner details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WinnerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No winner available
...
  /api/admin/games:
    get:
      tags: [Admin]
      operationId: getAllAdminGames
      description: Requires ADMIN role.
      responses:
        '200':
          description: All games for admin dashboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminGameDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/games/{gameId}:
    get:
      tags: [Admin]
      operationId: getAdminGameDetails
      description: Requires ADMIN role.
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGameDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin]
      operationId: deleteAdminGame
      description: Requires ADMIN role.
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/games/{gameId}/teams:
    get:
      tags: [Admin]
      operationId: getAdminGameTeams
      description: Requires ADMIN role.
      parameters:
        - $ref: '#/components/parameters/gameIdPath'
      responses:
        '200':
          description: Teams in game
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminTeamDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    gameIdPath:
      name: gameId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    teamIdPath:
      name: teamId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    roomIdPath:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    questionIdPath:
      name: questionId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    teamIdQuery:
      name: teamId
      in: query
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    QuestionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        image:
          type: string
          nullable: true
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerDto'
    RoomOfKnowledgeQuestionDto:
      type: object
      properties:
        question:
          type: string
        answers:
          type: array
          items:
            $ref: '#/components/schemas/RoomOfKnowledgeAnswerDto'
    RoomOfKnowledgeAnswerDto:
      type: object
      properties:
        answer:
          type: string
        isCorrect:
          type: boolean
    AnswerDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        text:
          type: string
    SubmitAnswerRequest:
      type: object
      required: [teamId, roomId, questionId, answerId]
      properties:
        teamId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
        questionId:
          type: integer
          format: int64
        answerId:
          type: integer
          format: int64
    SubmitAnswerResponse:
      type: object
      properties:
        correct:
          type: boolean
        pointsEarned:
          type: integer
          format: int32
        newTotalScore:
          type: integer
          format: int32
        feedback:
          type: string
        roomCompleted:
          type: boolean
        nextQuestionId:
          type: integer
          format: int64
          nullable: true
    RoomStatusResponse:
      type: object
      properties:
        roomId:
          type: integer
          format: int64
        location:
          $ref: '#/components/schemas/Location'
        status:
          $ref: '#/components/schemas/Status'
        timerRemaining:
          type: integer
          format: int32
        totalQuestions:
          type: integer
          format: int32
        answeredQuestions:
          type: integer
          format: int32
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionDto'
        isComplete:
          type: boolean
...
    GameResultDto:
      type: object
      properties:
        gameId:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          format: int64
        winnerId:
          type: integer
          format: int64
          nullable: true
        winnerName:
          type: string
          nullable: true
        teamResults:
          type: array
          items:
            $ref: '#/components/schemas/TeamResultDto'
    TeamResultDto:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        score:
          type: integer
          format: int32
        completionTimeMinutes:
          type: integer
          format: int32
        isWinner:
          type: boolean
        finalRanking:
          type: integer
          format: int32
    WinnerResponse:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        finalScore:
          type: integer
          format: int32
        gameDurationMinutes:
          type: integer
          format: int64
    AdminGameDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        password:
          type: string
        beginTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/Status'
        teamCount:
          type: integer
          format: int32
        activeTeams:
          type: integer
          format: int32
        completedTeams:
          type: integer
          format: int32
        teams:
          type: array
          items:
            $ref: '#/components/schemas/AdminTeamSummaryDto'
        winningTeamId:
          type: integer
          format: int64
          nullable: true
        winningTeamName:
          type: string
          nullable: true
    AdminTeamSummaryDto:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        mission:
          type: string
        status:
          $ref: '#/components/schemas/Status'
        score:
          type: integer
          format: int32
    AdminTeamDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mission:
          $ref: '#/components/schemas/Mission'
        status:
          $ref: '#/components/schemas/Status'
        currentLocation:
          $ref: '#/components/schemas/Location'
        score:
          type: integer
          format: int32
        questionsAnswered:
          type: integer
          format: int32
        questionsCorrect:
          type: integer
          format: int32
        hintsUsed:
          type: integer
          format: int32
        timeRemaining:
          type: integer
          format: int32
        startTime:
          type: string
          format: date-time
        completionTime:
          type: string
          format: date-time
          nullable: true
        gameId:
          type: integer
          format: int64
        gamePassword:
          type: string
        isWinner:
          type: boolean
        questionAttempts:
          type: array
          items:
            $ref: '#/components/schemas/AdminQuestionAttemptDto'
    AdminQuestionAttemptDto:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        questionTitle:
          type: string
        correct:
          type: boolean
        pointsEarned:
          type: integer
          format: int32
        hintUsed:
          type: boolean
        answeredAt:
          type: string
          format: date-time
...
    Location:
      type: string
      enum: [START, ROOM_OF_KNOWLEDGE, ROOM_OF_ABSTRACTS, ROOM_OF_ANALYTICS, ROOM_OF_SCIENCE_BATTLE, LAST_QUIZ, END_STATE]
    Mission:
      type: string
      enum:
        - WOUND_CARE_FOR_PRESSURE_ULCERS
        - FALL_PREVENTION_IN_GERIATRICS
        - PAIN_MANAGEMENT_IN_POSTOPERATIVE_CARE
        - NUTRITIONAL_INTERVENTIONS_FOR_MALNUTRITION
        - PREVENTION_OF_CATHETER_ASSOCIATED_URINARY_TRACT_INFECTIONS
    Status:
      type: string
      enum: [READY, STARTED, WAITING_APPROVAL, FINISHED]
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
      additionalProperties: true